Move connect to blocks
